Add iOS target for :data:db-sqldelight
