--- data/db-sqldelight/src/androidMain/kotlin/app/tivi/data/DriverFactory.kt
@@ -1,28 +1,0 @@
-// Copyright 2023, Christopher Banes and the Tivi project contributors
-// SPDX-License-Identifier: Apache-2.0
-
-package app.tivi.data
-
-import android.content.Context
-import androidx.sqlite.db.SupportSQLiteDatabase
-import app.cash.sqldelight.db.SqlDriver
-import app.cash.sqldelight.driver.android.AndroidSqliteDriver
-
-actual class DriverFactory(
-    private val context: Context,
-    private val databaseFilename: String,
-) {
-    actual fun createDriver(): SqlDriver {
-        return AndroidSqliteDriver(
-            schema = Database.Schema,
-            context = context,
-            name = databaseFilename,
-            callback = object : AndroidSqliteDriver.Callback(Database.Schema) {
-                override fun onConfigure(db: SupportSQLiteDatabase) {
-                    db.enableWriteAheadLogging()
-                    db.setForeignKeyConstraintsEnabled(true)
-                }
-            },
-        )
-    }
-}
