--- data/db-sqldelight/src/commonMain/kotlin/app/tivi/data/DriverFactory.kt
@@ -1,10 +1,0 @@
-// Copyright 2023, Christopher Banes and the Tivi project contributors
-// SPDX-License-Identifier: Apache-2.0
-
-package app.tivi.data
-
-import app.cash.sqldelight.db.SqlDriver
-
-expect class DriverFactory {
-    fun createDriver(): SqlDriver
-}
