+++ data/db-sqldelight/src/jvmMain/kotlin/app/tivi/data/SqlDelightDatabasePlatformComponent.kt
@@ -5,12 +5,15 @@ package app.tivi.data
 
 import app.cash.sqldelight.db.SqlDriver
 import app.cash.sqldelight.driver.jdbc.sqlite.JdbcSqliteDriver
+import app.tivi.inject.ApplicationScope
+import me.tatarka.inject.annotations.Provides
 
-actual class DriverFactory {
-    actual fun createDriver(): SqlDriver {
+actual interface SqlDelightDatabasePlatformComponent {
+    @Provides
+    @ApplicationScope
+    fun provideDriverFactory(): SqlDriver {
         return JdbcSqliteDriver(JdbcSqliteDriver.IN_MEMORY).also { db ->
             Database.Schema.create(db)
-
             db.execute(null, "PRAGMA foreign_keys=ON", 0)
         }
     }
